+++ modules/fluxcd/helm_operator.tf
@@ -23,9 +23,30 @@ resource "null_resource" "helm_operator_crd" {
   }
 }
 
-# bootstrap helm operator
+# bootstrap helm operator without values file
 resource "helm_release" "helm_operator" {
-  count = var.bootstrap_helm_operator ? 1 : 0
+  count = var.bootstrap_helm_operator && var.helm_operator_values_file == "" ? 1 : 0
+
+  depends_on = [
+    null_resource.helm_operator_crd[0]
+  ]
+
+  name       = var.helm_operator_spec_chart.release_name
+  repository = var.helm_operator_spec_chart.repository
+  chart      = var.helm_operator_spec_chart.chart_name
+  version    = var.helm_operator_spec_chart.chart_version
+
+  namespace = kubernetes_namespace.fluxcd.id
+
+  set {
+    name  = "git.secretName"
+    value = kubernetes_secret.fluxcd.metadata[0].name
+  }
+}
+
+# bootstrap helm operator with values file
+resource "helm_release" "helm_operator_w_values" {
+  count = var.bootstrap_helm_operator && var.helm_operator_values_file != "" ? 1 : 0
 
   depends_on = [
     null_resource.helm_operator_crd[0]
@@ -43,13 +64,7 @@ resource "helm_release" "helm_operator" {
     value = kubernetes_secret.fluxcd.metadata[0].name
   }
 
-  set {
-    name  = "helm.versions"
-    value = var.helm_operator_helm_versions
-  }
-
-  set {
-    name  = "prometheus.enabled"
-    value = var.helm_operator_prometheus_enabled
-  }
+  values = [
+    "${file(var.helm_operator_values_file)}"
+  ]
 }